.env updated

+++ database/migrations/2020_07_10_043610_create_smartcontracts_table.php
@@ -15,7 +15,7 @@ class CreateSmartcontractsTable extends Migration
     {
         Schema::create('smartcontracts', function (Blueprint $table) {
             $table->id();
-            $table->integer('userid');
+            $table->unsignedBigInteger('userid');
             $table->text('filename');
             $table->text('code');
             $table->foreign('userid')->references('id')->on('users');
